Completed doccumentation

+++ src/main/java/com/example/firstproject/TemperatureControl.java
@@ -1,8 +1,28 @@
 package com.example.firstproject;
 
+/**
+ * Интерфейс контроллера темрературы
+ */
 public interface TemperatureControl {
+    /**
+     * Включает контроллер
+     */
     void turnOn();
+
+    /**
+     * Выключает контроллер
+     */
     void turnOff();
-    void setTemperature(int temperature);
+
+    /**
+     * Задаёт нужную температуру
+     * @param temperature температура в градусах цельсия
+     */
+    void setTemperature(double temperature);
+
+    /**
+     * Получает текущую температуру
+     * @return температура в градусах цельсия
+     */
     double getTemperature();
 }
